bluedroid: Add option for a device-specific bluedroid implementation
We are in the process of adding ANT+ sensor support for tiwlan-based
SEMC devices. These binary plugs expect a slightly different bluedroid
implementation which may be useful for others implementing ANT+ but
may cause problems for other devices.

This patch adds BOARD_CUSTOM_BLUEDROID which can be used to replace the
standard bluetooth.c with a custom implementation without breaking
things for devices that assume default functionality.

Change-Id: I0f87d316d5c66a571626c4e3f0185ef51e69d07f